Add integration api test with Slim 4.
