+++ tests/integration/TestCase.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\integration;
 
-//use Exception;
 use PHPUnit\Framework\TestCase as PHPUnit_TestCase;
-//use Psr\Http\Message\ServerRequestInterface as Request;
 use Slim\Psr7\Factory\StreamFactory;
 use Slim\Psr7\Headers;
 use Slim\Psr7\Request as SlimRequest;
